add multi-field batch processing
Process prompts can declare an output.fields list so one structured LLM
request populates a related set of fields. Responses must contain exactly the
declared non-empty string values, and updates apply only after the complete
response validates.

Share the behavior across process-file and process-deck, preserve source and
undeclared fields, expose each changed field in previews and progress logs,
and keep retries, logging, resume, usage, and cost accounting on the existing
batch path. Deck processing skips partially populated target sets unless
--force is supplied and sends sparse atomic field updates to Anki.

Document the prompt syntax and safety behavior, retain output.field for
single-field prompts, and add focused coverage for validation, atomicity,
source-only data, resume, deck selection, progress, and structured requests.
The check-generated lint cleanups keep the full project checks warning-free.

+++ src/tts/provider/edge.rs
@@ -77,7 +77,7 @@ fn connect_request(muid: &str) -> Result<tungstenite::handshake::client::Request
     let url = format!(
         "{SYNTH_URL}{TRUSTED_CLIENT_TOKEN}&ConnectionId={connection_id}&Sec-MS-GEC={gec}&Sec-MS-GEC-Version={SEC_MS_GEC_VERSION}",
         connection_id = random_request_id(),
-        gec = generate_sec_ms_gec(now_unix_secs()? as u64),
+        gec = generate_sec_ms_gec(now_unix_secs()?),
     );
     let mut request = url
         .into_client_request()
